// src/interfaces/hero-interface.ts
export interface Hero {
  id: string;
  name: string;
  birth_year: string;
  eye_color: string;
  gender: string;
  hair_color: string;
  skin_color: string;
  homeworld: string;
  films: number[];
  species: number[];
  starships: number[];
  vehicles: number[];
  created: string;
  edited: string
  height: string;
  mass: string;
  url: string;
}